Updated to quit application gracefully.
Because CLI backend run in superuser permission, if we quit application
directly, it will not stop and become an orphan process. So we need to
tell it to stop manually. This can be done by destroy window manually
before application quits because we already handle CLI backend in window
dispose functions.